requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.9.5
async-timeout>=4.0.3
psutil>=5.9.5